add abstraction to process line to record for multi-record processor
