+++ kafka-connect-upload-source/src/main/java/org/radarbase/connect/upload/converter/altoida/AltoidaCsvMultiRecordsProcessor.kt
@@ -7,17 +7,27 @@ import org.radarbase.connect.upload.converter.CsvLineProcessorFactory
 import org.radarbase.connect.upload.converter.LogRepository
 import org.radarbase.connect.upload.converter.MultiRecordsCsvLineProcessor
 import org.slf4j.LoggerFactory
+import java.time.Instant
 
-abstract class AltoidaCsvMultiRecordsProcessor: CsvLineProcessorFactory {
+interface AltoidaSummaryLineToRecordMapper {
+    abstract val topic: String
+    fun time(line: Map<String, String>): Double = Instant.parse(line.getValue("TIMESTAMP")).toEpochMilli().toDouble()
+    fun processLine(line: Map<String, String>, timeReceived: Double): Pair<String, IndexedRecord>?
+}
+
+abstract class AltoidaCsvMultiRecordsProcessor : CsvLineProcessorFactory {
     private val logger = LoggerFactory.getLogger(javaClass)
 
     abstract val fileNameSuffix: String
 
-    abstract fun MultiRecordsCsvLineProcessor.lineConversion(line: Map<String, String>, timeReceived: Double): Map<String,IndexedRecord>?
+    abstract val listOfRecordMapperAltoidaSummaries: List<AltoidaSummaryLineToRecordMapper>
+
+    fun lineConversion(line: Map<String, String>, timeReceived: Double): Map<String, IndexedRecord>? =
+            listOfRecordMapperAltoidaSummaries.mapNotNull { it.processLine(line, timeReceived) }
+                    .asSequence()
+                    .toMap()
 
     override fun matches(contents: ContentsDTO): Boolean = contents.fileName.endsWith(fileNameSuffix)
-
-    open fun time(line: Map<String, String>): Double = line.getValue("TIMESTAMP").toDouble() / 1000.0
 
     override fun createLineProcessor(record: RecordDTO, logRepository: LogRepository): CsvLineProcessorFactory.CsvLineProcessor {
         val recordLogger = logRepository.createLogger(logger, record.id!!)
